fix(render): cast widened ptrCommandList at idpp assignment sites
ModelHeader.ptrCommandList is now uintptr_t; explicit (u32) cast at the
InstDrawPerPlayer copies keeps current behaviour. Full idpp pointer-field
widening is the M4 render-path work.

+++ game/RenderBucket/RenderBucket_QueueExecute.c
@@ -2028,7 +2028,7 @@ static struct RenderBucketEntry *RenderBucket_QueueDraw(struct Instance *inst, s
 	idpp->unkEC = drawFunc;
 	idpp->unkF0 = uncompressFunc;
 	RenderBucket_WriteInstanceCallbackLabels(inst, queuedFlags);
-	idpp->ptrCommandList = mh->ptrCommandList;
+	idpp->ptrCommandList = (u32)mh->ptrCommandList;
 	idpp->ptrTexLayout = mh->ptrTexLayout;
 	idpp->ptrColorLayout = (u32)mh->ptrColors;
 	idpp->instFlags = queuedFlags;
